requirements.txt
# Core data
pandas==2.2.2
numpy==1.26.4

# Machine learning
scikit-learn==1.5.1
# scikit-surprise requires C++ build tools via pip on Windows.
# Install with conda instead: conda install -c conda-forge scikit-surprise
# scikit-surprise==1.1.3
nltk==3.8.1

# Web app
streamlit==1.35.0
plotly==5.22.0

# Utilities
requests==2.32.3
python-dotenv==1.0.1
